+++ Packages/com.chark.game-management/Runtime/AssemblyInfo.cs
@@ -1,4 +1,5 @@
 ﻿using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("CHARK.GameManagement.Tests.Editor")]
+[assembly: InternalsVisibleTo("CHARK.GameManagement.Tests.Runtime")]
 [assembly: InternalsVisibleTo("CHARK.GameManagement.Editor")]